Integer not long

+++ src/main/org/jboss/invocation/jrmp/server/JRMPInvoker.java
@@ -54,7 +54,7 @@ import org.jboss.system.Registry;
 *
 *  @author <a href="mailto:marc.fleury@jboss.org>Marc Fleury</a>
 *
-*  @version $Revision: 1.7 $
+*  @version $Revision: 1.8 $
 *
 *  <p><b>Revisions:</b><br>
 *  <p><b>2002/01/13: Sacha Labourey</b>
@@ -316,7 +316,7 @@ implements Invoker, JRMPInvokerMBean,  MBeanRegistration
          // Extract the ObjectName, the rest is still marshalled
          // ObjectName mbean = new ObjectName((String) invocation.getContainer());
          
-         ObjectName mbean = (ObjectName) Registry.lookup((Long) invocation.getContainer());
+         ObjectName mbean = (ObjectName) Registry.lookup((Integer) invocation.getContainer());
          
          // The cl on the thread should be set in another interceptor
          return new MarshalledObject(server.invoke(mbean, invocation.getType(),  new Object[] {invocation}, new String[] {"java.lang.Object"}));
